Fix : index -> location.replace

+++ src/main/java/com/yejin/article/ArticleController.java
@@ -3,6 +3,7 @@ package com.yejin.article;
 import com.yejin.Rq;
 import com.yejin.article.dto.ArticleDto;
 
+
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -52,6 +53,8 @@ public class ArticleController {
         rq.appendBody("<div>title : %s</div>".formatted(title));
         rq.appendBody("<div>body : %s</div>".formatted(body));
         //rq.appendBody("<li><a href =\"/usr/article/detail/free/%d\">%d</a></li>".formatted(id,id));
+        rq.appendBody("<button><a href =\"/usr/article/detail/free/%d\">게시물 %d</a></button>".formatted(id,id));
+        rq.appendBody("<button><a href =\"/usr/article/list/free/\">게시물 목록</a></button>".formatted(id,id));
 
         showList(rq);
     }
@@ -96,7 +99,22 @@ public class ArticleController {
     }
 
     public void doDelete(Rq rq) {
+        System.out.println(rq.getPath());
         long id = rq.getLongPathValueByIndex(1,0);
+
+        if (id == 0) {
+            rq.appendBody("번호를 입력해주세요.");
+            return;
+        }
+
+        ArticleDto articleDto = articleService.findById(id);
+
+        if (articleDto == null) {
+            rq.appendBody("해당 글이 존재하지 않습니다.");
+            return;
+        }
+
+
         articleService.delete(id);
         rq.appendBody("<div>%d 번 게시물이 삭제되었습니다.</div>".formatted(id));
         // rq.appendBody("<div>title : %s</div>".formatted(title));
